field tables: rephrase and synthesize explanation
my second attempt at a revision of the field tables description! the
intent is to clarify that field tables describe both how information is
encoded (varints + byte sequences) but also that they define the exact
ordering of those bytes. that the fields are ordered is emphasized in
multiple ways in this revision to make it stick for readers. also have a
small hope a chunky intro paragraph signifies this section's importance
(so readers feel less inclined to skip the section!)

when i first read this section i took "Field Table" to be a term of art,
and only on revising the section did i grok that "oh yeah duh the name's
from being a table of fields". by breaking out that realization and
mentioning it explicitly i hope to make it easier for first-time readers
of this section to realise that what's happening is a bunch of fields
being ordered in a table (feels less intimidating than a capital noun
Field Table?)